fix 修复创建ssh失败问题，优化工作流增加状态

+++ script/package_windows.ps1
@@ -48,7 +48,9 @@ Copy-Item $WebExe (Join-Path $PackageDir "dtool.exe") -Force
 Copy-Item (Join-Path $RootDir "go.mod") (Join-Path $PackageDir "go.mod") -Force
 New-Item -ItemType Directory -Force -Path (Join-Path $PackageDir "config/dtool") | Out-Null
 Copy-Item (Join-Path $RootDir "config/dtool/company.ini") (Join-Path $PackageDir "config/dtool/config.ini") -Force
-Copy-Item (Join-Path $RootDir "config/dtool/frog.db") (Join-Path $PackageDir "config/dtool/frog.db") -Force
+if (Test-Path (Join-Path $RootDir "config/dtool/frog.db")) {
+    Copy-Item (Join-Path $RootDir "config/dtool/frog.db") (Join-Path $PackageDir "config/dtool/frog.db") -Force
+}
 Copy-Item $FrontendDistDir (Join-Path $PackageDir "web") -Recurse -Force
 Copy-Item (Join-Path $RootDir "internal/pkg/p_js") (Join-Path $PackageDir "internal/pkg/p_js") -Recurse -Force
 Copy-Item (Join-Path $RootDir "internal/app/dtool/database") (Join-Path $PackageDir "internal/app/dtool/database") -Recurse -Force
